refactor(engine): split templates into individual files
Extracts the monolithic src/engine/templates.ts (~510 lines) into a
src/engine/templates/ directory with types.ts, helpers.ts, default.ts,
social-card.ts, blog-hero.ts, email-banner.ts, and index.ts. The old
templates.ts becomes a barrel re-export so all existing imports remain
unchanged. Zero behavioral changes — all 168 tests pass.
